producer_service/requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.0
confluent-kafka==2.11.0
schedule==1.2.2
